Restyled components to match old design with new scss instead of chakra-ui
