+++ src/components/molecules/NewsArticleCard/NewsArticleCard.module.scss
@@ -1,5 +1,9 @@
 .newsArticle {
+    $width: 250px;
+
     background: lightgrey;
+    width: $width;
+    max-width: $width;
 
     p {
         padding: 5px;
@@ -7,8 +11,7 @@
     }
 
     img {
-        width: 250px;
-        height: 125px;
+        width: 100%;
         object-fit: fill;
     }
 }
