+++ src/components/organisms/Navbar/Navbar.module.scss
@@ -1,0 +1,42 @@
+@use 'src/styles/vars';
+@use 'src/styles/placeholders';
+
+.navbar {
+    $height: 100vh;
+    @extend %flex-column;
+
+    position: fixed;
+    background-color: vars.$color-chinese-white;
+    border-right: 1px;
+    width: vars.$navbar-width;
+    max-width: vars.$navbar-width;
+    height: $height;
+    max-height: $height;
+}
+
+.navHead {
+    font-size: 1.75rem;
+    border-bottom: 2px solid black;
+    text-align: center;
+    padding: 2rem;
+}
+
+.navLinks {
+    @extend %flex-column;
+    padding: 1rem;
+    gap: 0.75rem;
+}
+
+.navNews {
+    @extend %flex-column;
+    @extend %overflow-column;
+    padding: 1rem;
+    gap: 1rem;
+    flex-grow: 1;
+    align-items: flex-start;
+}
+
+.navNewsHead {
+    font-size: 1.6rem;
+    font-weight: 600;
+}
