+++ src/components/organisms/index.js
@@ -1,4 +1,4 @@
-export { Navbar } from './Navbar'
-export { ScrollCategory } from './ScrollCategory'
-export { NewsContainer } from './NewsContainer'
-export { GamesContainer } from './GamesContainer'
+export { Navbar } from './Navbar/Navbar'
+export { ScrollCategory } from './ScrollCategory/ScrollCategory'
+export { NewsContainer } from './NewsContainer/NewsContainer'
+export { GamesContainer } from './GamesContainer/GamesContainer'
